+++ src/aims_ui/page_singlesearch.py
@@ -5,7 +5,7 @@ from . import app
 from requests.exceptions import ConnectionError
 from .cookie_utils import save_input, load_input, get_all_inputs, delete_input, load_save_store_inputs, save_epoch_number
 from .api_interaction import api, get_response_attributes
-from .security_utils import detect_xml_injection
+from .security_utils import detect_xml_injection, check_user_has_access_to_page
 from .models.get_endpoints import get_endpoints
 from .models.get_fields import get_fields
 from .models.get_addresses import get_addresses
@@ -18,13 +18,17 @@ page_name = 'singlesearch'
 @login_required
 @app.route(f'/', methods=['GET', 'POST'])
 def singlesearch():
+  endpoints = get_endpoints(called_from=page_name)
+  access = check_user_has_access_to_page(page_name, endpoints)
+  if access != True:
+    return access
 
   if request.method == 'GET':
     delete_input(session)
     return render_template(
         f'{page_name}.html',
         searchable_fields=get_fields(page_name),
-        endpoints=get_endpoints(called_from=page_name),
+        endpoints=endpoints,
     )
 
   searchable_fields = get_fields(page_name)
@@ -79,7 +83,7 @@ def singlesearch():
 
   return render_template(
       f'{page_name}.html',
-      endpoints=get_endpoints(called_from=page_name),
+      endpoints=endpoints,
       searchable_fields=searchable_fields,
       results_page=True,
       matched_addresses=matched_addresses,
